Rename cache service to encoding  🐿 v2.5.14
