+++ test/server/modules/encoding/service.spec.js
@@ -1,9 +1,9 @@
 import sinon from 'sinon';
 import { expect } from 'chai';
 import proxyquire from 'proxyquire';
-import Cache from '../../../../server/modules/cache/service';
+import Encoder from '../../../../server/modules/encoding/service';
 
-describe('Cache Service', () => {
+describe('Encoder Service', () => {
 
 	const sandbox = sinon.sandbox.create();
 	let bufferStub;
@@ -30,7 +30,7 @@ describe('Cache Service', () => {
 
 		it('should convert data to a buffer', () => {
 			const cacheItem = { mock: 'object' };
-			Cache.encode(cacheItem);
+			Encoder.encode(cacheItem);
 
 			expect(bufferStub.calledOnce).to.eq(true);
 			expect(bufferStub.calledWith(JSON.stringify(cacheItem))).to.eq(true);
@@ -38,7 +38,7 @@ describe('Cache Service', () => {
 
 		it('should return a base64 string', () => {
 			const cacheItem = { mock: 'object' };
-			const encodedData = Cache.encode(cacheItem);
+			const encodedData = Encoder.encode(cacheItem);
 
 			expect(toStringStub.calledOnce).to.eq(true);
 			expect(toStringStub.calledWith('base64')).to.eq(true);
@@ -59,7 +59,7 @@ describe('Cache Service', () => {
 		});
 
 		it('should convert a base64 string to a buffer', () => {
-			Cache.decode(encodedStr);
+			Encoder.decode(encodedStr);
 
 			expect(bufferStub.calledOnce).to.eq(true);
 			expect(bufferStub.calledWith(encodedStr, 'base64')).to.eq(true);
@@ -67,13 +67,13 @@ describe('Cache Service', () => {
 		});
 
 		it('should parse and return the payload', () => {
-			const decodedData = Cache.decode(encodedStr);
+			const decodedData = Encoder.decode(encodedStr);
 			expect(decodedData).to.deep.equal(JSON.parse(toStringResponse));
 		});
 
 		it('should return null if the data is not valid', () => {
 			toStringStub.returns('THIS-IS-NOT-AN-OBJECT!')
-			const decodedData = Cache.decode(encodedStr);
+			const decodedData = Encoder.decode(encodedStr);
 			expect(decodedData).to.equal(null);
 		});
 
@@ -85,8 +85,8 @@ describe('Cache Service', () => {
 
 		it('should mean we actually confirm it works', () => {
 			const testData = { sanity: 'test' };
-			const encodedData = Cache.encode(testData);
-			const decodedData = Cache.decode(encodedData);
+			const encodedData = Encoder.encode(testData);
+			const decodedData = Encoder.decode(encodedData);
 
 			expect(decodedData).to.deep.equal(testData);
